Add files via upload
This is Macaulay2 code to iterate over the M2 database of smooth projective toric Fano 5-folds, apply the Hanlon-Hicks-Lazarev resolution of the diagonal to each 5-fold, and ask whether the collection of line bundles appearing on the left (i.e., either) side of the box product yields a full strong exceptional collection of line bundles. Comments are given throughout to explain the code.